+++ org.eclipse.scanning.test/META-INF/MANIFEST.MF
@@ -48,7 +48,7 @@ Require-Bundle: org.eclipse.scanning.api;bundle-version="1.0.0",
  org.eclipse.scanning.device;bundle-version="1.0.0",
  org.eclipse.richbeans.api;bundle-version="1.0.0",
  org.eclipse.richbeans.binding;bundle-version="1.0.0",
- org.eclipse.scanning.event.queues.spooler;bundle-version="1.0.0"
+ org.eclipse.scanning.event.queuesspooler;bundle-version="1.0.0"
 Service-Component: OSGI-INF/*.xml
 Import-Package: javax.jms,
  org.apache.commons.lang;version="2.6.0",
